# app/docs/swagger.yaml
basePath: /v1
definitions:
  model.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  model.PLayList:
    properties:
      description:
        example: A short description of the application
        type: string
      level:
        example: 34
        type: integer
      title:
        example: Title name
        type: string
    type: object
...
  model.PlaylistTracksResponse:
    properties:
      playlist:
        $ref: '#/definitions/model.PLayList'
      tracks:
        items:
          $ref: '#/definitions/model.Track'
        type: array
    type: object
  model.PlaylistsResponse:
    properties:
      playlists:
        items:
          $ref: '#/definitions/model.PLayList'
        type: array
    type: object
...
  model.Track:
    properties:
      album:
        example: Album name
        type: string
      album_artist:
        example: Album artist name
        type: string
      artist:
        example: Artist name
        type: string
      bitrate:
        example: 320
        type: integer
      comment:
        example: Additional comments
        type: string
      composer:
        example: Composer name
        type: string
      disc:
        example: 1
        type: integer
      disc_total:
        example: 2
        type: integer
      duration:
        $ref: '#/definitions/time.Duration'
      genre:
        example: Genre name
        type: string
      lyrics:
        example: Lyrics of the track
        type: string
      sample_rate:
        example: 44100
        type: integer
      title:
        example: Title name
        type: string
      track:
        example: 3
        type: integer
      track_total:
        example: 10
        type: integer
      year:
        example: 2022
        type: integer
    type: object
...
  time.Duration:
    enum:
    - -9223372036854775808
    - 9223372036854775807
    - 1
    - 1000
    - 1000000
    - 1000000000
    - 60000000000
    - 3600000000000
    - -9223372036854775808
    - 9223372036854775807
    - 1
    - 1000
    - 1000000
    - 1000000000
    - 60000000000
    - 3600000000000
    - 1
    - 1000
    - 1000000
    type: integer
    x-enum-varnames:
    - minDuration
    - maxDuration
    - Nanosecond
    - Microsecond
    - Millisecond
    - Second
    - Minute
    - Hour
    - minDuration
    - maxDuration
    - Nanosecond
    - Microsecond
    - Millisecond
    - Second
    - Minute
    - Hour
    - Nanosecond
    - Microsecond
    - Millisecond
...
host: s3streammedia.com
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a s3 media streamer server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: S3 Media Streamer Application API
  version: 0.0.1
paths:
...
  /playlist/{playlist_id}:
    delete:
      consumes:
      - application/json
      description: Delete a playlist based on its unique ID.
      parameters:
      - description: Playlist ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a playlist by ID.
      tags:
      - playlist-controller
    get:
      consumes:
      - application/json
      description: Get tracks from a playlist by providing the playlist ID.
      parameters:
      - description: Playlist ID
        in: path
        name: playlist_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Tracks retrieved successfully
          schema:
            $ref: '#/definitions/model.PlaylistTracksResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get tracks from a playlist.
      tags:
      - playlist-controller
    post:
      consumes:
      - application/json
      description: Set tracks in a playlist by providing a list of track_handler IDs.
      parameters:
      - description: Playlist ID
        in: path
        name: playlist_id
        required: true
        type: string
      - description: List of track_handler IDs to set in the playlist
        in: body
        name: track_ids
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: Tracks set in the playlist successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Set tracks in a playlist.
      tags:
      - playlist-controller
  /playlist/{playlist_id}/{track_id}:
    delete:
      consumes:
      - application/json
      description: Remove a track_handler from the specified playlist.
      parameters:
      - description: Playlist ID
        in: path
        name: playlist_id
        required: true
        type: string
      - description: Track ID
        in: path
        name: track_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Track removed from playlist successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Playlist or track_handler not found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Remove a track_handler from the playlist.
      tags:
      - playlist-controller
    post:
      consumes:
      - application/json
      description: Add a track_handler to an existing playlist.
      parameters:
      - description: Playlist ID
        in: path
        name: playlist_id
        required: true
        type: string
      - description: Track ID
        in: path
        name: track_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Track added to the playlist successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Playlist or track_handler not found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Add a track_handler to a playlist
      tags:
      - playlist-controller
  /playlist/{playlist_id}/clear:
    delete:
      consumes:
      - application/json
      description: Removes all tracks from a playlist, effectively clearing it.
      parameters:
      - description: Playlist ID
        in: path
        name: playlist_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Playlist cleared successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Playlist not found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Clear a playlist by removing all tracks from it.
      tags:
      - playlist-controller
  /playlist/create:
    post:
      consumes:
      - application/json
      description: Creates a new playlist with the provided information.
      parameters:
      - description: PLayList details
        in: body
        name: request
        required: true
        schema:
          items:
            $ref: '#/definitions/model.PLayList'
          type: array
      produces:
      - application/json
      responses:
        "201":
          description: Playlist created successfully
          schema:
            items:
              $ref: '#/definitions/model.PLayList'
            type: array
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized - User unauthenticated
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new playlist.
      tags:
      - playlist-controller
  /playlist/get:
    get:
      consumes:
      - application/json
      description: Retrieves all playlists available in the storage.
      produces:
      - application/json
      responses:
        "200":
          description: Playlists retrieved successfully
          schema:
            $ref: '#/definitions/model.PlaylistsResponse'
        "404":
          description: Playlists not found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all playlists
      tags:
      - playlist-controller
...
schemes:
- http
- https
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"